implemented separate component for add member functionality
